+++ neard/Cargo.toml
@@ -63,7 +63,7 @@ delay_detector = ["nearcore/delay_detector"]
 rosetta_rpc = ["nearcore/rosetta_rpc"]
 json_rpc = ["nearcore/json_rpc"]
 protocol_feature_fix_staking_threshold = ["nearcore/protocol_feature_fix_staking_threshold"]
-protocol_feature_flat_state = ["nearcore/protocol_feature_flat_state"]
+protocol_feature_flat_state = ["nearcore/protocol_feature_flat_state", "near-flat-storage/protocol_feature_flat_state"]
 protocol_feature_nep366_delegate_action = [
   "nearcore/protocol_feature_nep366_delegate_action",
   "near-state-viewer/protocol_feature_nep366_delegate_action",
@@ -73,6 +73,7 @@ nightly = [
   "nightly_protocol",
   "nearcore/nightly",
   "near-state-viewer/nightly",
+  "protocol_feature_flat_state",
 ]
 nightly_protocol = ["nearcore/nightly_protocol"]
