+++ tools/flat-storage/Cargo.toml
@@ -9,23 +9,19 @@ edition.workspace = true
 anyhow.workspace = true
 borsh.workspace = true
 clap.workspace = true
-num-rational.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-tempfile.workspace = true
-tracing.workspace = true
 rayon.workspace = true
 
 tqdm = "0.4.4"
 
 near-chain = { path = "../../chain/chain" }
 near-chain-configs = { path = "../../core/chain-configs" }
-near-crypto = { path = "../../core/crypto" }
+
 near-epoch-manager = { path = "../../chain/epoch-manager" }
-near-network = { path = "../../chain/network" }
 near-primitives = { path = "../../core/primitives" }
-near-primitives-core = { path = "../../core/primitives-core" }
 near-store = { path = "../../core/store" }
-near-test-contracts = { path = "../../runtime/near-test-contracts" }
 nearcore = { path = "../../nearcore" }
-node-runtime = { path = "../../runtime/runtime" }
+
+[features]
+protocol_feature_flat_state = ["nearcore/protocol_feature_flat_state",
+"near-chain/protocol_feature_flat_state",
+"near-store/protocol_feature_flat_state",]
